refactor: update imports for new LabelGroup component location
